requirements.txt
flask
python-dotenv
requests
cohere